Navigation in tables now managed with NavigationBtns component
