+++ api-hooks/inventory-api/getItemsPerPage.ts
@@ -10,7 +10,7 @@ export const getItemsPerPage = (token: string, page: number) => {
 				await getInventoryTemplate(token).get('/tableRegisters', {
 					params: {
 						page,
-						size: 10,
+						size: 1,
 						sort: []
 					},
 				})
