+++ components/common/TableInventory.tsx
@@ -2,15 +2,13 @@
 import React, { useState } from 'react';
 import { TableRowInventory } from './table/TableRowInventory';
 import { Table } from './table/Table';
-import NextPrevButton from './table/NextPrevButton';
 import { getItemsTableHeaders } from '@/api-hooks/inventory-api/getItemTableHeaders';
 import { useSession } from 'next-auth/react';
 import { getItemsPerPage } from '@/api-hooks/inventory-api/getItemsPerPage';
-import { useQueryClient } from '@tanstack/react-query';
+import NavigationBtns from './table/NavigationBtns';
 
 const TableInventory = () => {
 	const token = useSession().data?.token?.token;
-	const queryClient = useQueryClient();
 	const [currentPage, setCurrentPage] = useState<number>(0);
 
 	const {
@@ -26,27 +24,6 @@ const TableInventory = () => {
 	} = getItemsPerPage(token as string, currentPage);
 
 	const totalPages = itemsPage?.totalPages;
-
-	const goToNextPage = () => {
-		if (totalPages !== undefined && currentPage < totalPages - 1) {
-			setCurrentPage((prev) => prev + 1);
-			queryClient
-				.refetchQueries({ queryKey: ['items-per-page'] })
-				.catch((error) => {
-					console.error('Error invalidating queries:', error);
-				});
-		}
-	};
-	const goToPreviousPage = () => {
-		if (currentPage > 0) {
-			setCurrentPage((prev) => prev - 1);
-			queryClient
-				.refetchQueries({ queryKey: ['items-per-page'] })
-				.catch((error) => {
-					console.error('Error invalidating queries:', error);
-				});
-		}
-	};
 
 	if (isErrorHeaders) return <p>Header Error: {errorHeaders.message}</p>;
 	if (isErrorItems) return <p>Item Error: {errorItems.message}</p>;
@@ -82,21 +59,12 @@ const TableInventory = () => {
 				</div>
 
 				<div className='flex items-center mt-4 gap-x-4 sm:mt-0'>
-					<NextPrevButton
-						onClick={goToPreviousPage}
-						disabled={currentPage === 0}
-						text={'Anterior'}
-						d={'M6.75 15.75L3 12m0 0l3.75-3.75M3 12h18'}
-						svg={'http://www.w3.org/2000/svg'}
-					></NextPrevButton>
-
-					<NextPrevButton
-						onClick={goToNextPage}
-						disabled={currentPage + 1 === totalPages}
-						text={'Siguiente'}
-						d={'M17.25 8.25L21 12m0 0l-3.75 3.75M21 12H3'}
-						svg={'http://www.w3.org/2000/svg'}
-					></NextPrevButton>
+					<NavigationBtns
+						currentPage={currentPage}
+						queryKey='items-per-page'
+						setCurrentPage={setCurrentPage}
+						totalPages={totalPages as number}
+					/>
 				</div>
 			</div>
 		</div>
