+++ components/common/TableLoan.tsx
@@ -2,16 +2,13 @@
 import React, { useState } from 'react';
 import { TableRowLoan } from './table/TableRowLoan';
 import { Table } from './table/Table';
-import NextPrevButton from './table/NextPrevButton';
 import { getLoanPerPage } from '@/api-hooks/loan-api/getLoanPerPageQuery';
 import { useSession } from 'next-auth/react';
 import { getLoanTableHeaders } from '@/api-hooks/loan-api/getLoanTableHeaders';
-import { useQueryClient } from '@tanstack/react-query';
-
+import NavigationBtns from './table/NavigationBtns';
 
 const TableLoan = () => {
 	const token = useSession().data?.token?.token;
-	const queryClient = useQueryClient();
 	const [currentPage, setCurrentPage] = useState<number>(0);
 
 	const {
@@ -28,26 +25,6 @@ const TableLoan = () => {
 
 	const totalPages = loanPage?.totalPages;
 
-	const goToNextPage = () => {
-		if (totalPages !== undefined && currentPage < totalPages - 1) {
-			setCurrentPage((prev) => prev + 1);
-			queryClient
-				.refetchQueries({ queryKey: ['loan-per-page'] })
-				.catch((error) => {
-					console.error('Error invalidating queries:', error);
-				});
-		}
-	};
-	const goToPreviousPage = () => {
-		if (currentPage > 0) {
-			setCurrentPage((prev) => prev - 1);
-			queryClient
-				.refetchQueries({ queryKey: ['loan-per-page'] })
-				.catch((error) => {
-					console.error('Error invalidating queries:', error);
-				});
-		}
-	};
 	if (isErrorHeaders)
 		return <p>Error cargando los headers: {errorHeaders.message}</p>;
 	if (isErrorLoan)
@@ -79,23 +56,12 @@ const TableLoan = () => {
 					</span>
 				</div>
 
-				<div className='flex items-center mt-4 gap-x-4 sm:mt-0'>
-					<NextPrevButton
-						onClick={goToPreviousPage}
-						disabled={currentPage === 0}
-						text={'Anterior'}
-						d={'M6.75 15.75L3 12m0 0l3.75-3.75M3 12h18'}
-						svg={'http://www.w3.org/2000/svg'}
-					></NextPrevButton>
-
-					<NextPrevButton
-						onClick={goToNextPage}
-						disabled={currentPage + 1 === totalPages}
-						text={'Siguiente'}
-						d={'M17.25 8.25L21 12m0 0l-3.75 3.75M21 12H3'}
-						svg={'http://www.w3.org/2000/svg'}
-					></NextPrevButton>
-				</div>
+				<NavigationBtns
+					totalPages={totalPages as number}
+					setCurrentPage={setCurrentPage}
+					currentPage={currentPage}
+					queryKey='loan-per-page'
+				/>
 			</div>
 		</div>
 	);
